Deprecating IdentityUser and rename IdentityUserV2 and other V2 named objects. Lots of cleanup.
