+++ src/ElCamino.AspNetCore.Identity.AzureTable/Model/IdentityUser.cs
@@ -10,11 +10,11 @@ using Microsoft.Azure.Cosmos.Table;
 
 namespace ElCamino.AspNetCore.Identity.AzureTable.Model
 {
-    public class IdentityUserV2 : IdentityUser<string>, IGenerateKeys
+    public class IdentityUser : IdentityUser<string>, IGenerateKeys
     {
-        public IdentityUserV2() : base() { }
+        public IdentityUser() : base() { }
 
-        public IdentityUserV2(string userName)
+        public IdentityUser(string userName)
             : this()
         {
             this.UserName = userName;
@@ -60,91 +60,6 @@ namespace ElCamino.AspNetCore.Identity.AzureTable.Model
                 }
             }
         }
-    }
-
-    [Obsolete("IdentityUserV2 will be renamed IdentityUser in a future release. Use IdentityUserV2 instead to ease breaking changes.")]
-    public class IdentityUser : IdentityUser<string, IdentityUserLogin, IdentityUserRole, IdentityUserClaim, IdentityUserToken>, IGenerateKeys
-    {
-        public IdentityUser() : base() { }
-
-        public IdentityUser(string userName)
-            : this()
-        {
-            this.UserName = userName;
-        }
-
-        /// <summary>
-        /// Generates Row, Partition and Id keys.
-        /// All are the same in this case
-        /// </summary>
-        public void GenerateKeys()
-        {
-            if (string.IsNullOrWhiteSpace(Id))
-            {
-                Id = KeyHelper.GenerateUserId();
-            }
-            RowKey = PeekRowKey();
-            PartitionKey = RowKey;
-            KeyVersion = KeyHelper.KeyVersion;
-        }
-
-        /// <summary>
-        /// Generates the RowKey without setting it on the object.
-        /// In this case, just returns a key based on username
-        /// </summary>
-        /// <returns></returns>
-        public string PeekRowKey()
-        {
-            return KeyHelper.GenerateRowKeyUserId(Id);
-        }
-
-        public double KeyVersion { get; set; }
-
-
-        public override string UserName
-        {
-            get => base.UserName;
-            set
-            {
-                if (!string.IsNullOrWhiteSpace(value))
-                {
-                    base.UserName = value.Trim();
-                }
-            }
-        }
-    }
-
-    [Obsolete("IdentityUser<TKey, TLogin, TRole, TClaim, TToken> : IdentityUser<TKey> will be removed in a future release.")]
-    public class IdentityUser<TKey, TLogin, TRole, TClaim, TToken> : IdentityUser<TKey>
-        where TKey : IEquatable<TKey>
-        where TLogin : IdentityUserLogin<TKey>
-        where TRole : IdentityUserRole<TKey>
-        where TClaim : IdentityUserClaim<TKey>
-        where TToken : IdentityUserToken<TKey>
-    {
-        public IdentityUser()
-        {
-            this.Claims = new List<TClaim>(10);
-            this.Roles = new List<TRole>(10);
-            this.Logins = new List<TLogin>(10);
-            this.Tokens = new List<TToken>(10);
-        }
-
-        #region Collections
-        [IgnoreProperty]
-        public ICollection<TClaim> Claims { get; private set; }
-
-        [IgnoreProperty]
-        public ICollection<TLogin> Logins { get; private set; }
-
-        [IgnoreProperty]
-        public ICollection<TRole> Roles { get; private set; }
-
-        [IgnoreProperty]
-        public ICollection<TToken> Tokens { get; private set; }
-
-        #endregion
-
     }
 
     public class IdentityUser<TKey> : Microsoft.AspNetCore.Identity.IdentityUser<TKey>, ITableEntity
