+++ src/ElCamino.AspNetCore.Identity.AzureTable/RoleStore.cs
