+++ tests/ElCamino.AspNetCore.Identity.AzureTable.Tests/ModelTests/ApplicationUser.cs
@@ -4,15 +4,7 @@ using ElCamino.AspNetCore.Identity.AzureTable.Model;
 
 namespace ElCamino.Web.Identity.AzureTable.Tests.ModelTests
 {
-#pragma warning disable 0618
-    public class ApplicationUser : IdentityUser, IApplicationUser
-    {
-        public string FirstName { get; set; }
-
-        public string LastName { get; set; }
-    }
-#pragma warning restore 0618
-    public class ApplicationUserV2 : IdentityUserV2, IApplicationUser
+    public class ApplicationUserV2 : IdentityUser, IApplicationUser
     {
         public string FirstName { get; set; }
