+++ tests/ElCamino.AspNetCore.Identity.AzureTable.Tests/RoleStoreTests.cs
@@ -15,12 +15,11 @@ using Microsoft.Extensions.DependencyInjection;
 
 namespace ElCamino.AspNetCore.Identity.AzureTable.Tests
 {
-#pragma warning disable 0618
-    public class RoleStoreTests : IClassFixture<RoleFixture<IdentityUserV2, IdentityRole, IdentityCloudContext>>
+    public class RoleStoreTests : IClassFixture<RoleFixture<Model.IdentityUser, IdentityRole, IdentityCloudContext>>
     {
         private readonly ITestOutputHelper output;
-        private RoleFixture<IdentityUserV2, IdentityRole, IdentityCloudContext> roleFixture;
-        public RoleStoreTests(RoleFixture<IdentityUserV2, IdentityRole, IdentityCloudContext> roleFix, ITestOutputHelper output)
+        private RoleFixture<Model.IdentityUser, IdentityRole, IdentityCloudContext> roleFixture;
+        public RoleStoreTests(RoleFixture<Model.IdentityUser, IdentityRole, IdentityCloudContext> roleFix, ITestOutputHelper output)
         {
             this.output = output;
             roleFixture = roleFix;
@@ -29,7 +28,6 @@ namespace ElCamino.AspNetCore.Identity.AzureTable.Tests
             }).Wait();
             
         }
-#pragma warning restore 0618
 
         [Fact(DisplayName = "RoleStoreCtors")]
         [Trait("IdentityCore.Azure.RoleStore", "")]
@@ -143,14 +141,12 @@ namespace ElCamino.AspNetCore.Identity.AzureTable.Tests
 
             ServiceCollection services = new ServiceCollection();
             // Adding coverage for CreateAzureTablesIfNotExists();
-#pragma warning disable 0618
             services.AddIdentityCore<IdentityUser>()
-                .AddAzureTableStoresV2<IdentityCloudContext>(new Func<IdentityConfiguration>(() =>
+                .AddAzureTableStores<IdentityCloudContext>(new Func<IdentityConfiguration>(() =>
                 {
                     return roleFixture.GetConfig();
                 }))
                 .CreateAzureTablesIfNotExists<IdentityCloudContext>();
-#pragma warning restore 0618
 
         }
