+++ tests/ElCamino.AspNetCore.Identity.AzureTable.Tests/UserStoreTests.cs
@@ -11,7 +11,8 @@ using Microsoft.AspNetCore.Builder;
 using Xunit;
 using Xunit.Abstractions;
 using ElCamino.AspNetCore.Identity.AzureTable;
-using IdentityUser = ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityUser<string>;
+using ElCamino.AspNetCore.Identity.AzureTable.Model;
+using IdentityUser = ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityUser;
 using IdentityRole = ElCamino.AspNetCore.Identity.AzureTable.Model.IdentityRole;
 using ElCamino.Web.Identity.AzureTable.Tests.ModelTests;
 using ElCamino.Web.Identity.AzureTable.Tests.Fixtures;
@@ -19,84 +20,90 @@ using Microsoft.AspNetCore.Identity;
 
 namespace ElCamino.AspNetCore.Identity.AzureTable.Tests
 {
-#pragma warning disable 0618
-    public partial class UserStoreTests : BaseUserStoreTests<ApplicationUserV2, IdentityCloudContext, UserOnlyStore<ApplicationUserV2, IdentityCloudContext>>
+    public partial class UserStoreTests : BaseUserStoreTests<ApplicationUserV2, IdentityRole, IdentityCloudContext, UserStore<ApplicationUserV2, IdentityRole, IdentityCloudContext>>
     {
-        public const string UserOnlyStoreTrait = "IdentityCore.Azure.UserOnlyStore";
-        public UserStoreTests(UserFixture<ApplicationUserV2, IdentityCloudContext, UserOnlyStore<ApplicationUserV2, IdentityCloudContext>> userFix, ITestOutputHelper output) :
-            base(userFix, output) {  }
-#pragma warning restore 0618
+        public const string UserStoreTrait = "IdentityCore.Azure.UserStore";
+        public const string UserStoreTraitProperties = UserStoreTrait + ".Properties";
+
+        public UserStoreTests(
+            UserFixture<ApplicationUserV2, IdentityRole, IdentityCloudContext,
+                UserStore<ApplicationUserV2, IdentityRole, IdentityCloudContext>> userFix, ITestOutputHelper output) :
+            base(userFix, output)
+        {
+            
+        }
+
         [Fact(DisplayName = "AddRemoveUserClaim")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task AddRemoveUserClaim()
         {
             return base.AddRemoveUserClaim();
         }
 
         [Fact(DisplayName = "AddRemoveUserLogin")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task AddRemoveUserLogin()
         {
             return base.AddRemoveUserLogin();
         }
 
-        //[Fact(DisplayName = "AddRemoveUserRole")]
-        //[Trait(UserOnlyStoreTrait, "")]
-        //public override Task AddRemoveUserRole()
-        //{
-        //    return base.AddRemoveUserRole();
-        //}
+        [Fact(DisplayName = "AddRemoveUserRole")]
+        [Trait(UserStoreTrait, "")]
+        public override Task AddRemoveUserRole()
+        {
+            return base.AddRemoveUserRole();
+        }
 
         [Fact(DisplayName = "AddRemoveUserToken")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task AddRemoveUserToken()
         {
             return base.AddRemoveUserToken();
         }
 
         [Fact(DisplayName = "AddReplaceRemoveUserClaim")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task AddReplaceRemoveUserClaim()
         {
             return base.AddReplaceRemoveUserClaim();
-        }
+        } 
 
         [Fact(DisplayName = "AddUserClaim")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task AddUserClaim()
         {
             return base.AddUserClaim();
         }
 
         [Fact(DisplayName = "AddUserLogin")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task AddUserLogin()
         {
             return base.AddUserLogin();
         }
 
-        //[Fact(DisplayName = "AddUserRole")]
-        //[Trait(UserOnlyStoreTrait, "")]
-        //public override Task AddUserRole()
-        //{
-        //    return base.AddUserRole();
-        //}
+        [Fact(DisplayName = "AddUserRole")]
+        [Trait(UserStoreTrait, "")]
+        public override Task AddUserRole()
+        {
+            return base.AddUserRole();
+        }
 
         [Fact(DisplayName = "ChangeUserName")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task ChangeUserName()
         {
             return base.ChangeUserName();
         }
 
         [Fact(DisplayName = "CheckDupEmail")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task CheckDupEmail()
         {
             return base.CheckDupEmail();
         }
 
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         [Fact(DisplayName = "CheckDupUser")]
         public override Task CheckDupUser()
         {
@@ -104,107 +111,246 @@ namespace ElCamino.AspNetCore.Identity.AzureTable.Tests
         }
 
         [Fact(DisplayName = "CreateUser")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task CreateUserTest()
         {
             return base.CreateUserTest();
         }
 
         [Fact(DisplayName = "DeleteUser")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task DeleteUser()
         {
             return base.DeleteUser();
         }
 
         [Fact(DisplayName = "FindUserByEmail")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task FindUserByEmail()
         {
             return base.FindUserByEmail();
         }
 
         [Fact(DisplayName = "FindUserById")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task FindUserById()
         {
             return base.FindUserById();
         }
 
         [Fact(DisplayName = "FindUserByName")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task FindUserByName()
         {
             return base.FindUserByName();
         }
 
         [Fact(DisplayName = "FindUsersByEmail")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task FindUsersByEmail()
         {
             return base.FindUsersByEmail();
         }
 
         //[Fact(DisplayName = "GenerateUsers", Skip = "true")]
-        //[Trait(UserOnlyStoreTrait, "")]
+        //[Trait(UserStoreTrait, "")]
         //public override Task GenerateUsers()
         //{
         //    return base.GenerateUsers();
         //}
 
         [Fact(DisplayName = "GetUsersByClaim")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task GetUsersByClaim()
         {
             return base.GetUsersByClaim();
         }
 
-        //[Fact(DisplayName = "GetUsersByRole")]
-        //[Trait(UserOnlyStoreTrait, "")]
-        //public override Task GetUsersByRole()
-        //{
-        //    return base.GetUsersByRole();
-        //}
+        [Fact(DisplayName = "GetUsersByRole")]
+        [Trait(UserStoreTrait, "")]
+        public override Task GetUsersByRole()
+        {
+            return base.GetUsersByRole();
+        }
 
-        //[Fact(DisplayName = "IsUserInRole")]
-        //[Trait(UserOnlyStoreTrait, "")]
-        //public override Task IsUserInRole()
-        //{
-        //    return base.IsUserInRole();
-        //}
+        [Fact(DisplayName = "IsUserInRole")]
+        [Trait(UserStoreTrait, "")]
+        public override Task IsUserInRole()
+        {
+            return base.IsUserInRole();
+        }
 
         [Fact(DisplayName = "ThrowIfDisposed")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task ThrowIfDisposed()
         {
             return base.ThrowIfDisposed();
         }
 
         [Fact(DisplayName = "UpdateApplicationUser")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task UpdateApplicationUser()
         {
             return base.UpdateApplicationUser();
         }
 
         [Fact(DisplayName = "UpdateUser")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override Task UpdateUser()
         {
             return base.UpdateUser();
         }
 
-#pragma warning disable 0618
         [Fact(DisplayName = "UserStoreCtors")]
-        [Trait(UserOnlyStoreTrait, "")]
+        [Trait(UserStoreTrait, "")]
         public override void UserStoreCtors()
         {
-            Assert.Throws<ArgumentNullException>(() => 
+            Assert.Throws<ArgumentNullException>(() =>
             {
-                new UserOnlyStore<ApplicationUser, IdentityCloudContext>(null,null);
+                new UserStore<ApplicationUserV2, IdentityRole, IdentityCloudContext>(null,null);
             });
         }
-#pragma warning restore 0618
-    }
 
+        #region Properties
+
+        [Fact(DisplayName = "AccessFailedCount")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task AccessFailedCount()
+        {
+            return base.AccessFailedCount();
+        }
+
+        [Fact(DisplayName = "Email")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task Email()
+        {
+            return base.Email();
+        }
+
+        [Fact(DisplayName = "EmailConfirmed")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task EmailConfirmed()
+        {
+            return base.EmailConfirmed();
+        }
+
+        [Fact(DisplayName = "EmailNone")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task EmailNone()
+        {
+            return base.EmailNone();
+        }
+
+        [Fact(DisplayName = "LockoutEnabled")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task LockoutEnabled()
+        {
+            return base.LockoutEnabled();
+        }
+
+        [Fact(DisplayName = "PasswordHash")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task PasswordHash()
+        {
+            return base.PasswordHash();
+        }
+
+        [Fact(DisplayName = "PhoneNumber")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task PhoneNumber()
+        {
+            return base.PhoneNumber();
+        }
+
+        [Fact(DisplayName = "PhoneNumberConfirmed")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task PhoneNumberConfirmed()
+        {
+            return base.PhoneNumberConfirmed();
+        }
+
+        [Fact(DisplayName = "SecurityStamp")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task SecurityStamp()
+        {
+            return base.SecurityStamp();
+        }
+
+        [Fact(DisplayName = "TwoFactorEnabled")]
+        [Trait(UserStoreTraitProperties, "")]
+        public override Task TwoFactorEnabled()
+        {
+            return base.TwoFactorEnabled();
+        }
+
+        #endregion
+
+
+        [Fact(DisplayName = "UserIdNotChangedIfImmutableIdSetUp")]
+        [Trait(UserStoreTraitProperties, "")]
+
+        public async Task UserIdNotChangedIfImmutableIdSetUp()
+        {
+            var userStore = GetImmutableUserIdStore();
+
+            var user = GenTestUser();
+            await userStore.CreateAsync(user);
+
+            var idBefore = user.Id;
+            var pkBefore = user.PartitionKey;
+            var rkBefore = user.RowKey;
+
+            user.UserName += "changed";
+            await userStore.UpdateAsync(user);
+
+            Assert.Equal(idBefore, user.Id);
+            Assert.Equal(pkBefore, user.PartitionKey);
+            Assert.Equal(rkBefore, user.RowKey);
+
+        }
+
+        private UserStore<ApplicationUserV2, IdentityRole, IdentityCloudContext> GetImmutableUserIdStore()
+        {
+            var config = userFixture.GetConfig();
+            var userStore = userFixture.CreateUserStore(userFixture.GetContext(config), config);
+            return userStore;
+        }
+
+        
+        [Fact(DisplayName = "CanFindByNameIfImmutableIdSetUp")]
+        [Trait(UserStoreTraitProperties, "")]
+
+        public async Task CanFindByNameIfImmutableIdSetUp()
+        {
+            var userStore = GetImmutableUserIdStore();
+
+            var user = GenTestUser();
+            await userStore.CreateAsync(user);
+
+            var userFound = await userStore.FindByNameAsync(user.UserName);
+
+            Assert.NotNull(user);
+            Assert.Equal(user.Id, userFound.Id);
+            Assert.Equal(user.PartitionKey, userFound.PartitionKey);
+            Assert.Equal(user.RowKey, userFound.RowKey);
+        }
+
+        [Fact(DisplayName = "CanFindByIdIfImmutableIdSetUp")]
+        [Trait(UserStoreTraitProperties, "")]
+
+        public async Task CanFindByIdIfImmutableIdSetUp()
+        {
+            var userStore = GetImmutableUserIdStore();
+
+            var user = GenTestUser();
+            await userStore.CreateAsync(user);
+
+            var userFound = await userStore.FindByIdAsync(user.Id);
+
+            Assert.NotNull(user);
+            Assert.Equal(user.Id, userFound.Id);
+            Assert.Equal(user.PartitionKey, userFound.PartitionKey);
+            Assert.Equal(user.RowKey, userFound.RowKey);
+        }
+    }
 }
